backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.3.0

# HTTP client (for LLM APIs)
httpx==0.26.0

# Excel export
openpyxl==3.1.2

# Utilities
python-dateutil==2.8.2

# Financial calculations
numpy==1.26.3
numpy-financial==1.0.0

# Document extraction
PyMuPDF==1.23.8
Pillow==10.2.0
scipy==1.12.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
httpx==0.26.0
